Add files via upload
This repository contains the description and solutions for a data science assignment. The assignment covers various topics in machine learning and statistical analysis. Below is an outline of the questions addressed in the assignment:

Overview of ensemble techniques in machine learning.
Reasons for using ensemble techniques and their advantages.
Explanation of bagging and boosting techniques.
Benefits of using ensemble techniques.
Comparison between ensemble techniques and individual models.
Calculation of confidence intervals using bootstrap resampling.
Description of the bootstrap resampling method and its steps.
Application of bootstrap resampling to estimate confidence intervals.
The assignment also includes a specific problem involving the estimation of the mean height of a population of trees using bootstrap resampling