Add files via upload
his repository contains a comprehensive guide to understanding and implementing the K-Nearest Neighbors (KNN) algorithm in machine learning. KNN is a simple yet powerful supervised learning algorithm used for both classification and regression tasks. In this guide, we explore the fundamental concepts of KNN and address various aspects related to its application